0529PC

+++ Assets/Script/Common/Tower/TowerTemplate.cs
@@ -14,7 +14,7 @@ public class TowerTemplate : ScriptableObject //타워 레벨 디자인
         public float attack;
         public float attackIncrease; // 공격력증가량
         public int cost; //업그레이드 비용
-        public int probability; //업그레이드 성공확률
+        public float probability; //업그레이드 성공확률
         public int fail;//업그레이드 실패확률
         public int sell; //타워 판매 시 획득 골드
     }
